Clarify reserves on CFTs
Free CFTs should not be supported because they go against the rationale of having ledger reserves. Ledger reserves exist to protect the node operators from having to store free data on behalf of ledger users, without some cost. Making a "free tier" for CFTs goes against that idea, and shouldn't be modified on a per-feature basis because reserve fees aren't really related to any particular single feature. Instead, they're a ledge-wide solution to a sort of ledger-wide economic problem.

Instead, if we want to propose a fee change, we should propose that idea independent of any particular ledger feature.

Last but not least, this free tier doesn’t seem to be necessary because a user must “opt in” to hold CFTs anyway, so that user can decide at that point if they want to pay the 2XRP reserve or not (this isn't a very large amount to bear anyway).