Clarify TransferFee Rounding (#226)
In the implementation of MPTs, TransferFees are rounded to the nearest whole value if the fraction is less than 0.5, and rounded up to the nearest whole value if the fraction is greater than or equal to 0.5. This provides the fairest rounding allocation between issuer and token holder. This change corrects the spec to align with the chosen implementation.